Move theming settings validation registration into ready method
This avoids premature registration of django checks which could break
django 1.11 startup.
